# services/assistant-service/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisRequest": {
        "properties": {
          "max_llm_calls": {
            "default": 20,
            "title": "Max Llm Calls",
            "type": "integer"
          },
          "max_sql_executions": {
            "default": 5,
            "title": "Max Sql Executions",
            "type": "integer"
          },
          "max_steps": {
            "default": 10,
            "title": "Max Steps",
            "type": "integer"
          },
          "max_wall_time_sec": {
            "default": 180,
            "title": "Max Wall Time Sec",
            "type": "integer"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "threadId": {
            "title": "Threadid",
            "type": "string"
          }
        },
        "required": [
          "question",
          "threadId"
        ],
        "title": "AnalysisRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Intelligent Assistant Orchestration System",
    "title": "Pixels Rover Assistant Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analysis": {
      "post": {
        "description": "Submit an analysis question. Returns SSE stream of events.",
        "operationId": "submit_analysis_api_v1_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Analysis",
        "tags": [
          "analysis"
        ]
      }
    },
...
    "/api/v1/analysis/{session_id}": {
      "get": {
        "description": "Get a completed analysis result by session ID.",
        "operationId": "get_analysis_result_api_v1_analysis__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analysis Result",
        "tags": [
          "analysis"
        ]
      }
    },
...
  }
}